requirements.txt
requests==2.32.3
PyYAML==6.0.2
tldextract==5.3.0
python-dateutil==2.9.0.post0
warcio==1.7.4         # used only if you later switch to true recording mode